use build job names again for TARGET/ARCH
GitLab CI seem to lose some variables when creating downstream jobs,
revert the changes back to use the job name for target/arch and apply
the rule to SDK testing.

Remove leading `deploy` from build jobs so they are better readable in
the GitLab UI. Use the leading build name as TYPE variable, so that SDKs
are not labeled on target base and imagebuilder/rootfs is not labeled
based on arch.

For `generate_targets.sh`, instead of exporting BRANCH as env variable,
use functions.sh to set it automatically. Export the `VERSION` one
globally instead of for each downstream job. This works after removing
the `gitlab-ci.yml` variable which would previously overwrite
(strangely) the downstream job settings.

Signed-off-by: Paul Spooren <mail@aparcar.org>