ci: s/UPSTREAM_URL/FILE_HOST and use mirror-03 as upstream provider
The job is triggered whenever a new target is uploaded. This happens the same
moment the CDN cache is flushed, so it's possible that there is a flaky
situation where the CDN is half way through purging the cache.

Unify the variables for the upstream URL, in Dockerfile it's called FILE_HOST,
in the setup.sh it's called UPSTREAM_URL. Use FILE_HOST everywhere but fallback
to UPSTREAM_URL if defined.

Let's use our own mirror for this.

Signed-off-by: Paul Spooren <mail@aparcar.org>